+++ students/jtschelling/go.mod
@@ -1,0 +1,5 @@
+module github.com/jtschelling/urlshort/students/jtschelling
+
+go 1.15
+
+require github.com/tidwall/buntdb v1.1.2
